pyproject.toml
[project]
name = "deepin-mcp-servers"
version = "0.1.0"
description = "Deepin MCP Server implementation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "requests>=2.30.0",
    "PyPDF2",
    "python-docx",
    "pandas",
    "python-pptx",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "markdownify>=0.11.0",
    "readabilipy>=0.2.0",
    "readability-lxml>=0.8.1",
    "lxml_html_clean"
]
